move util functions and remove comments
